Refactor BlockListItem for better structure.
